requirements.txt
gevent
webapp2
Jinja2
python-memcached
peewee
ws4py